+++ src/components/PerfList.tsx
@@ -1,15 +1,6 @@
-import { useContext } from 'react';
-
-import { Button, Flex, Icon, VStack, Wrap, WrapItem } from '@chakra-ui/react';
-import { RiArrowLeftLine, RiDownloadLine } from 'react-icons/ri';
-import { useLocation } from 'wouter';
+import { VStack, Wrap, WrapItem } from '@chakra-ui/react';
 
 import { SerieT } from '@/modules/series/SeriesTypes';
-
-import useDownloadPerfCsv from '@/hooks/useDownloadPerfCsv';
-import { PerfBrowserContext } from '@/hooks/usePerfBrowser';
-
-import { compilePath, PATHS } from '@/routes';
 
 import PerfCard from '@/components/PerfCard';
 import PerfChart from '@/components/PerfChart';
@@ -20,9 +11,6 @@ interface PerfListProps {
     onCardClick: (metricName: string) => void;
 }
 const PerfList = ({ seriesGroups, onCardClick }: PerfListProps) => {
-    const { computePlans } = useContext(PerfBrowserContext);
-    const downloadPerfCsv = useDownloadPerfCsv(seriesGroups);
-    const [, setLocation] = useLocation();
     return (
         <VStack
             flexGrow={1}
@@ -30,39 +18,9 @@ const PerfList = ({ seriesGroups, onCardClick }: PerfListProps) => {
             justifyContent="flex-start"
             alignItems="stretch"
             spacing="4"
-            paddingX="8"
-            paddingY="4"
+            padding="8"
             overflow="hidden"
         >
-            <Flex justifyContent="space-between" alignItems="center">
-                <Button
-                    leftIcon={<RiArrowLeftLine />}
-                    onClick={() =>
-                        setLocation(
-                            computePlans.length > 1
-                                ? PATHS.COMPUTE_PLANS
-                                : compilePath(PATHS.COMPUTE_PLAN_TASKS_ROOT, {
-                                      key: computePlans[0].key,
-                                  })
-                        )
-                    }
-                    variant="ghost"
-                    size="md"
-                    fontWeight="medium"
-                >
-                    Go back
-                </Button>
-                <Button
-                    leftIcon={<Icon as={RiDownloadLine} />}
-                    variant="solid"
-                    colorScheme="teal"
-                    size="sm"
-                    onClick={downloadPerfCsv}
-                    isDisabled={seriesGroups.length === 0}
-                >
-                    Download as CSV
-                </Button>
-            </Flex>
             <PerfEmptyState seriesGroups={seriesGroups} />
             <Wrap spacing="3">
                 {seriesGroups.map((series) => (